init kendra
