+++ query/query.http
@@ -3,7 +3,7 @@ Content-Type: application/json
 
 {
   "IndexId": "65a73c7f-74df-43da-950b-70cbc46be778",
-  "QueryText": "kendra"
+  "QueryText": "who is sung?"
 }
